Refactor `ReactorUtils` into its own `sentry-reactor` module
