+++ sentry-reactor/api/sentry-reactor.api
@@ -1,0 +1,26 @@
+public final class io/sentry/reactor/BuildConfig {
+	public static final field SENTRY_REACTOR_SDK_NAME Ljava/lang/String;
+	public static final field VERSION_NAME Ljava/lang/String;
+}
+
+public final class io/sentry/reactor/SentryReactorThreadLocalAccessor : io/micrometer/context/ThreadLocalAccessor {
+	public static final field KEY Ljava/lang/String;
+	public fun <init> ()V
+	public fun getValue ()Lio/sentry/IScopes;
+	public synthetic fun getValue ()Ljava/lang/Object;
+	public fun key ()Ljava/lang/Object;
+	public fun reset ()V
+	public fun setValue (Lio/sentry/IScopes;)V
+	public synthetic fun setValue (Ljava/lang/Object;)V
+}
+
+public final class io/sentry/reactor/SentryReactorUtils {
+	public fun <init> ()V
+	public static fun withSentry (Lreactor/core/publisher/Flux;)Lreactor/core/publisher/Flux;
+	public static fun withSentry (Lreactor/core/publisher/Mono;)Lreactor/core/publisher/Mono;
+	public static fun withSentryForkedRoots (Lreactor/core/publisher/Flux;)Lreactor/core/publisher/Flux;
+	public static fun withSentryForkedRoots (Lreactor/core/publisher/Mono;)Lreactor/core/publisher/Mono;
+	public static fun withSentryScopes (Lreactor/core/publisher/Flux;Lio/sentry/IScopes;)Lreactor/core/publisher/Flux;
+	public static fun withSentryScopes (Lreactor/core/publisher/Mono;Lio/sentry/IScopes;)Lreactor/core/publisher/Mono;
+}
+
