+++ sentry-samples/sentry-samples-spring-boot-jakarta-opentelemetry-noagent/build.gradle.kts
@@ -53,6 +53,7 @@ dependencies {
     implementation(projects.sentryGraphql22)
     implementation(projects.sentryQuartz)
     implementation(projects.sentryOpentelemetry.sentryOpentelemetryAgentlessSpring)
+    implementation(projects.sentryReactor)
 
     // database query tracing
     implementation(projects.sentryJdbc)
