+++ sentry-samples/sentry-samples-spring-boot-jakarta-opentelemetry/build.gradle.kts
@@ -54,6 +54,7 @@ dependencies {
     implementation(projects.sentryGraphql22)
     implementation(projects.sentryQuartz)
     implementation(Config.Libs.OpenTelemetry.otelSdk)
+    implementation(projects.sentryReactor)
 
     // database query tracing
     implementation(projects.sentryJdbc)
