+++ sentry-samples/sentry-samples-spring-boot-jakarta/build.gradle.kts
@@ -52,6 +52,7 @@ dependencies {
     implementation(projects.sentryLogback)
     implementation(projects.sentryGraphql22)
     implementation(projects.sentryQuartz)
+    implementation(projects.sentryReactor)
 
     // database query tracing
     implementation(projects.sentryJdbc)
