+++ sentry-spring-jakarta/src/main/resources/META-INF/services/io.micrometer.context.ThreadLocalAccessor
@@ -1,1 +1,1 @@
-io.sentry.spring.jakarta.webflux.SentryReactorThreadLocalAccessor
+io.sentry.reactor.SentryReactorThreadLocalAccessor
